debug velocity params saving
