+++ dynamo/tools/utils.py
@@ -26,13 +26,10 @@ from tqdm import tqdm
 
 from ..dynamo_logger import (
     LoggerManager,
-    main_critical,
     main_debug,
-    main_exception,
     main_info,
     main_info_insert_adata,
     main_info_verbose_timeit,
-    main_tqdm,
     main_warning,
 )
 from ..preprocessing.transform import _Freeman_Tukey
@@ -1607,22 +1604,6 @@ def set_param_kinetic(
     adata.var = var
 
     return adata
-
-
-def convert_velocity_params_type(adata: AnnData) -> None:
-    """Convert the velocity parameters into numeric type to avoid error when saving to h5ad.
-
-    The data type of the parameters in the Pandas Series is set to object due to the initialization of parameters with
-    None values. However, this choice of data type can lead to errors during the process of saving the anndata object.
-    Therefore, it becomes necessary to convert these parameters to a numeric data type to ensure proper handling and
-    compatibility.
-    """
-    velocity_param_names = ["beta", "gamma", "half_life", "alpha_b", "alpha_r2", "gamma_b", "gamma_r2", "gamma_logLL",
-                            "delta_b", "delta_r2", "bs", "bf", "uu0", "ul0", "su0", "sl0", "alpha", "a", "b", "alpha_a",
-                            "alpha_a", "alpha_i", "cost", "logLL"]
-    for param_name in velocity_param_names:
-        if param_name in adata.var.columns:
-            adata.var[param_name] = pd.to_numeric(adata.var[param_name], errors='coerce')
 
 
 def get_U_S_for_velocity_estimation(subset_adata, use_moments, has_splicing, has_labeling, log_unnormalized, NTR):
